26-03-2014, matex 1. Fixed a bug in ASC and ASCPAD with negative hex values

+++ 0.56.b/common/CExpression.cpp
@@ -516,7 +516,7 @@ try_dec:
 						if ( pszArgs && *pszArgs )
 						{
 							iCount = 1;
-							iResult = static_cast<int>(exp(static_cast<double>(GetVal(pszArgs))));
+							iResult = static_cast<INT64>(exp(static_cast<double>(GetVal(pszArgs))));
 						}
 						else
 						{
